Integration tests for AppAreaService added. AppAreaService updated.

+++ src/main/java/com/workfront/internship/workflow/service/impl/AppAreaServiceImpl.java
@@ -1,11 +1,10 @@
 package com.workfront.internship.workflow.service.impl;
 
 import com.workfront.internship.workflow.dao.AppAreaDAO;
-import com.workfront.internship.workflow.dao.impl.AppAreaDAOImpl;
-import com.workfront.internship.workflow.dao.impl.PostDAOImpl;
 import com.workfront.internship.workflow.domain.AppArea;
 import com.workfront.internship.workflow.domain.User;
 import com.workfront.internship.workflow.exceptions.service.InvalidObjectException;
+import com.workfront.internship.workflow.exceptions.service.ServiceLayerException;
 import com.workfront.internship.workflow.service.AppAreaService;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,7 +17,7 @@ import java.util.List;
  */
 public class AppAreaServiceImpl implements AppAreaService {
 
-    private static final Logger logger = Logger.getLogger(PostDAOImpl.class);
+    private static final Logger LOGGER = Logger.getLogger(AppAreaServiceImpl.class);
 
     private final AppAreaDAO appAreaDAO;
 
@@ -29,31 +28,65 @@ public class AppAreaServiceImpl implements AppAreaService {
 
     @Override
     public long add(AppArea appArea) {
-        if (!appArea.isValid()){
-            logger.error("Application Area is invalid. Failed to add to the database");
+        if (appArea == null || !appArea.isValid()) {
+            LOGGER.error("Application Area is invalid. Failed to add to the database");
             throw new InvalidObjectException("Invalid Application Area");
         }
         long id = 0;
         try{
             id = appAreaDAO.add(appArea);
         }catch (RuntimeException e){
-            logger.error("Failed to add the application area to database");
+            LOGGER.error("Failed to add the application area to database");
         }
         return id;
     }
 
     @Override
-    public List<User> getUsersById(long appAreaId) {
-        return appAreaDAO.getUsersById(appAreaId);
+    public List<User> getUsersById(long id) {
+        if (id < 1) {
+            LOGGER.error("Id is not valid");
+            throw new InvalidObjectException("Not valid id");
+        }
+
+        List<User> users;
+        try {
+            users = appAreaDAO.getUsersById(id);
+        } catch (RuntimeException e) {
+            LOGGER.error("Failed to delete the user");
+            throw new ServiceLayerException("Failed to delete the user", e);
+        }
+        return users;
     }
 
     @Override
     public AppArea getById(long id) {
-        return appAreaDAO.getById(id);
+        if (id < 1) {
+            LOGGER.error("Id is not valid");
+            throw new InvalidObjectException("Not valid id");
+        }
+
+        AppArea appArea;
+        try {
+            appArea = appAreaDAO.getById(id);
+        } catch (RuntimeException e) {
+            LOGGER.error("Failed to delete the user");
+            throw new ServiceLayerException("Failed to delete the user", e);
+        }
+        return appArea;
     }
 
     @Override
     public void deleteById(long id) {
-        appAreaDAO.deleteById(id);
+        if (id < 1) {
+            LOGGER.error("Id is not valid");
+            throw new InvalidObjectException("Not valid id");
+        }
+
+        try {
+            appAreaDAO.deleteById(id);
+        } catch (RuntimeException e) {
+            LOGGER.error("Failed to delete the user");
+            throw new ServiceLayerException("Failed to delete the user", e);
+        }
     }
 }
